Query the Github API for a list of issues associated with a repository

The issues for a given repository and all of their metadata are returned as a tibble. The table can then be used to summarize information about a given repository such as how many open issues there are. Information about closed issues is not returned.

get_issues(org = "NOAA-FIMS", repo = "FIMS", from = -Inf)

Arguments

org

The name of the organization that contains the repo. This can also be a user name rather than an organization. The default is "NOAA-FIMS".

repo

The name of the GitHub repository. The default is "FIMS".

from

The date that you would like to start the activity from. The default is negative infinity, which leads to all entries being returned. But, another viable option is Sys.Date() - 1, which provides everything since yesterday. Note that some entries prior to this entry will also be returned if they are included in the same page, as the entire page is returned.

Value

A 32-column tibble, that was created from the json list that is returned from gh::gh(), is returned. The returned tibble contains the following columns:

url

repository_url

labels_url

comments_url

events_url

html_url

id

node_id

number

title

user

labels

state

locked

assignee

assignees

milestone

comments

created_at

updated_at

closed_at

author_association

sub_issues_summary

active_lock_reason

draft

pull_request: A five-column data frame with information about the Pull Request if the issue was a actually a Pull Request and not an issue. All entries are NULL if it was in fact an issue. Columns include url, html_url, diff_url, patch_url, and merged_at.

body

closed_by

reactions

timeline_url

performed_via_github_app

state_reason

Examples

if (FALSE) { # \dontrun{
get_issues(repo = "FIMS", org = "NOAA-FIMS", from = Sys.Date() - 1)
} # }