Get a list of repositories for a given user

Get a list of repositories for a given user

get_repositories_user(user)

Arguments

user

A string providing the GitHub user name of interest.

Value

A tibble is returned with the following three columns:

user: A string with the provided GitHub user name.

name: A string giving the name of the repository.

fork: A logical, specifying if the repository is a fork.

Examples

get_repositories_user("kellijohnson-NOAA")
#> # A tibble: 26 × 3
#>    user              name                        fork 
#>    <chr>             <chr>                       <lgl>
#>  1 kellijohnson-NOAA .github                     FALSE
#>  2 kellijohnson-NOAA alaska                      FALSE
#>  3 kellijohnson-NOAA AR-perf-testing             FALSE
#>  4 kellijohnson-NOAA BrailleR                    TRUE 
#>  5 kellijohnson-NOAA deploy-bookdown             TRUE 
#>  6 kellijohnson-NOAA Dirichlet-Multinomial       TRUE 
#>  7 kellijohnson-NOAA FishLife                    TRUE 
#>  8 kellijohnson-NOAA git-references              FALSE
#>  9 kellijohnson-NOAA iwccla                      FALSE
#> 10 kellijohnson-NOAA kellijohnson-NOAA.github.io FALSE
#> # ℹ 16 more rows
data <- purrr::map_df(
  c("kellijohnson-NOAA", "msupernaw"),
  get_repositories_user
)