Builds and sends the email

This function compiles the tables with some email text and sends it using blastula::smtp_send.

Usage

build_email(from_, to_, creds, body_, table_iss, table_pr, footer_)

Arguments

from_

sender email address

to_

recipient email address

creds

A credential file of the type supported by codeblastula

body_

The text to precede the table

table_iss

the issue table

table_pr

The PR table

footer_

the text to follow the table

Value

the email body

See also